+++ RTF++/src/RTD/Strings/SharedString.cpp
@@ -61,13 +61,13 @@ namespace RTF {
         SharedString SharedString::operator+(const SharedString &string) {
             SharedString str;
             str.string.SetSize(this->Size() + string.Size() + 1);
-            uint64_t i = 0;
+            size_t i = 0;
 
             for(; i < this->string.Size(); i++){
                 str.string[i] = this->string[i];
             }
 
-            for(uint64_t j = 0; j <= string.Size(); i++, j++){
+            for(size_t j = 0; j <= string.Size(); i++, j++){
                 str.string[i] = string.string.GetElement(j);
             }
 
@@ -76,16 +76,16 @@ namespace RTF {
 
         SharedString SharedString::operator+(const char *string1) {
             SharedString str;
-            uint64_t size = strlen(string1) + 1;
+            size_t size = strlen(string1) + 1;
 
             str.string.SetSize(this->Size() + size);
-            uint64_t i = 0;
+            size_t i = 0;
 
             for(; i < this->string.Size(); i++){
                 str.string[i] = this->string[i];
             }
 
-            for(uint64_t j = 0; j <= size; i++, j++){
+            for(size_t j = 0; j <= size; i++, j++){
                 str.string[i] = string1[j];
             }
 
@@ -110,7 +110,7 @@ namespace RTF {
         bool SharedString::operator==(const SharedString& string) {
             if(this->Size() != string.Size())
                 return false;
-            for(uint64_t i = 0; i < this->string.Size(); i++)
+            for(size_t i = 0; i < this->string.Size(); i++)
                 if((*this->string)[i] != (*((Containers::SharedVector<char>&)string.string))[i])
                     return false;
 
@@ -119,7 +119,7 @@ namespace RTF {
         bool SharedString::operator==(const char* str) {
             if(this->Size() != strlen(str))
                 return false;
-            for(uint64_t i = 0; i < this->string.Size(); i++)
+            for(size_t i = 0; i < this->string.Size(); i++)
                 if((*this->string)[i] != str[i])
                     return false;
 
@@ -134,7 +134,7 @@ namespace RTF {
             return !(*this == str);
         }
 
-        char& SharedString::operator[](uint64_t i){
+        char& SharedString::operator[](size_t i){
             return this->string[i];
         }
 
@@ -154,7 +154,7 @@ namespace RTF {
             return *this;
         }
 
-        uint64_t SharedString::Size() const{
+        size_t SharedString::Size() const{
             return this->string.Size() - 1;
         }
